+++ evaluator/src/commonTest/kotlin/com/github/jcornaz/kwik/evaluator/ForAll1Test.kt
@@ -1,7 +1,6 @@
 package com.github.jcornaz.kwik.evaluator
 
 import com.github.jcornaz.kwik.generator.api.Generator
-import com.github.jcornaz.kwik.generator.api.withSamples
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -13,21 +12,6 @@ class ForAll1Test : AbstractRunnerTest() {
 
     override fun evaluate(iterations: Int, seed: Long, invocation: PropertyEvaluationContext.() -> Boolean) {
         forAll(testGenerator, iterations, seed) { invocation() }
-    }
-
-    @Test
-    fun evaluateSamples() {
-        val values = mutableSetOf<Int>()
-
-        val gen = Generator.create { it.nextInt(0, 10) }.withSamples(42, 100)
-
-        forAll<Int>(gen) {
-            values += it
-            true
-        }
-
-        assertTrue(42 in values)
-        assertTrue(100 in values)
     }
 
     @Test
@@ -59,7 +43,7 @@ class ForAll1Test : AbstractRunnerTest() {
             Generator.create { ++iterations },
             iterations = 10
         ) { x ->
-            ensureAtLeastOne("is greater than 100") { x >= 100 }
+            ensureAtLeastOne { x >= 100 }
             true
         }
 
@@ -74,8 +58,24 @@ class ForAll1Test : AbstractRunnerTest() {
             Generator.create { ++iterations },
             iterations = 10
         ) { x ->
-            ensureAtLeastOne("is greater than 100") { x >= 100 }
-            ensureAtLeastOne("is greater than 10") { x >= 10 }
+            ensureAtLeastOne { x >= 100 }
+            ensureAtLeastOne { x >= 10 }
+            true
+        }
+
+        assertEquals(100, iterations)
+    }
+
+    @Test
+    fun multipleEnsureAtLeastOneCauseAdditionalIterationUntilTheyAreBothSatisfied_orderDoesNotMatter() {
+        var iterations = 0
+
+        forAll(
+            Generator.create { ++iterations },
+            iterations = 10
+        ) { x ->
+            ensureAtLeastOne { x >= 10 }
+            ensureAtLeastOne { x >= 100 }
             true
         }
 
@@ -91,7 +91,7 @@ class ForAll1Test : AbstractRunnerTest() {
             iterations = 123
         ) { x ->
             ++iteration
-            ensureAtLeastOne("is greated than 10") { x > 10 }
+            ensureAtLeastOne { x > 10 }
             true
         }
 
@@ -108,7 +108,7 @@ class ForAll1Test : AbstractRunnerTest() {
                 seed = 78
             ) { x ->
                 ++iteration
-                ensureAtLeastOne("is greater than 10") { x > 10 }
+                ensureAtLeastOne { x > 10 }
                 iteration < 10
             }
         }