+++ generator/api/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/api/GeneratorCombination.kt
@@ -2,8 +2,6 @@ package com.github.jcornaz.kwik.generator.api
 
 import kotlin.random.Random
 
-private const val NUMBER_OF_SAMPLES_FOR_COMBINATION = 5
-
 /**
  * Returns a generator of values built from the elements of `this` generator and the [other] generator
  * using the provided [transform] function applied to each pair of elements.
@@ -12,7 +10,7 @@ fun <A, B, R> Generator<A>.combineWith(
     other: Generator<B>,
     transform: (A, B) -> R
 ): Generator<R> =
-    CombinedGenerators(this, other, transform)
+    Generator.combine(this, other, transform)
 
 /**
  * Returns a generator of values built from the elements of `this` generator and the [other] generator
@@ -20,7 +18,7 @@ fun <A, B, R> Generator<A>.combineWith(
 fun <A, B> Generator<A>.combineWith(
     other: Generator<B>
 ): Generator<Pair<A, B>> =
-    CombinedGenerators(this, other, ::Pair)
+    Generator.combine(this, other, ::Pair)
 
 /**
  * Returns a generator of combining the elements of [generator1] and [generator2]
@@ -40,18 +38,13 @@ fun <A, B> Generator.Companion.combine(
     generator1: Generator<A>,
     generator2: Generator<B>
 ): Generator<Pair<A, B>> =
-    CombinedGenerators(generator1, generator2, ::Pair)
+    combine(generator1, generator2, ::Pair)
 
 private class CombinedGenerators<A, B, R>(
     private val generator1: Generator<A>,
     private val generator2: Generator<B>,
     private val transform: (A, B) -> R
 ) : Generator<R> {
-    override val samples: Set<R> = generator1.samples.take(NUMBER_OF_SAMPLES_FOR_COMBINATION)
-        .flatMapTo(mutableSetOf()) { a ->
-            generator2.samples.take(NUMBER_OF_SAMPLES_FOR_COMBINATION).map { b -> transform(a, b) }
-        }
-
     override fun generate(random: Random): R =
         transform(generator1.generate(random), generator2.generate(random))
 }
@@ -66,8 +59,6 @@ private class MergedGenerators<T>(
     private val generator1: Generator<T>,
     private val generator2: Generator<T>
 ) : Generator<T> {
-    override val samples: Set<T> = generator1.samples + generator2.samples
-
     override fun generate(random: Random): T =
         if (random.nextBoolean()) generator1.generate(random) else generator2.generate(random)
 }
@@ -92,6 +83,7 @@ fun <T> Generator.Companion.frequency(
     val list = weightedGenerators.asSequence()
         .onEach { (weight, _) -> require(weight >= 0.0) { "Negative weight(s) found in frequency input" } }
         .filter { (weight, _) -> weight > 0.0 }
+        .sortedByDescending { (weight, _) -> weight }
         .toList()
 
     return when (list.size) {
@@ -119,24 +111,19 @@ fun <T> Generator.Companion.frequency(
  * @throws IllegalArgumentException if [weightedGenerators] is empty, contains negative weights or the sum of the weight is zero
  */
 fun <T> Generator.Companion.frequency(
-    vararg weightedGenerator: Pair<Double, Generator<T>>
-): Generator<T> = frequency(weightedGenerator.asList())
+    vararg weightedGenerators: Pair<Double, Generator<T>>
+): Generator<T> = frequency(weightedGenerators.asList())
 
 private class DualGenerator<T>(
     private val source1: Generator<T>,
     private val source2: Generator<T>,
     private val source1Probability: Double
 ) : Generator<T> {
-    override val samples: Set<T> = source1.samples + source2.samples
-
     override fun generate(random: Random): T =
         if (random.nextDouble() < source1Probability) source1.generate(random) else source2.generate(random)
 }
 
 private class FrequencyGenerator<T>(private val weightedGenerators: List<Pair<Double, Generator<T>>>) : Generator<T> {
-    override val samples: Set<T> = weightedGenerators
-        .flatMapTo(HashSet()) { (_, gen) -> gen.samples }
-
     val max = weightedGenerators.sumByDouble { (weight, _) -> weight }
 
     init {