+++ generator/api/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/api/WithSampleTest.kt
@@ -1,10 +1,7 @@
 package com.github.jcornaz.kwik.generator.api
 
 import com.github.jcornaz.kwik.generator.test.AbstractGeneratorTest
-import kotlin.test.Test
-import kotlin.test.assertEquals
-import kotlin.test.assertSame
-import kotlin.test.assertTrue
+import kotlin.test.*
 
 class WithSampleTest : AbstractGeneratorTest() {
 
@@ -13,17 +10,52 @@ class WithSampleTest : AbstractGeneratorTest() {
             .withSamples(1, 2, 3, 4)
 
     @Test
-    fun hasSamples() {
-        val generator = Generator.create { it.nextInt(5, Int.MAX_VALUE) }
-            .withSamples(1, 2, 3, 4)
-
-        assertEquals(setOf(1, 2, 3, 4), generator.samples)
-    }
-
-    @Test
     fun emptyListOfSamplesReturnOriginalGenerator() {
         val gen = Generator.create { it.nextInt() }
         assertSame(gen, gen.withSamples())
+    }
+
+    @Test
+    fun failsIfProbabilityIsBellowZero() {
+        val source = Generator.create { it.nextInt(5, Int.MAX_VALUE) }
+        assertFailsWith<IllegalArgumentException> {
+            source.withSamples(1, 2, 3, 4, probability = -1.0)
+        }
+    }
+
+    @Test
+    fun failsIfProbabilityIsZero() {
+        val source = Generator.create { it.nextInt(5, Int.MAX_VALUE) }
+        assertFailsWith<IllegalArgumentException> {
+            source.withSamples(1, 2, 3, 4, probability = 0.0)
+        }
+    }
+
+    @Test
+    fun failsIfProbabilityOne() {
+        val source = Generator.create { it.nextInt(5, Int.MAX_VALUE) }
+        assertFailsWith<IllegalArgumentException> {
+            source.withSamples(1, 2, 3, 4, probability = 1.0)
+        }
+    }
+
+    @Test
+    fun failsIfProbabilityAboveOne() {
+        val source = Generator.create { it.nextInt(5, Int.MAX_VALUE) }
+        assertFailsWith<IllegalArgumentException> {
+            source.withSamples(1, 2, 3, 4, probability = 1.01)
+        }
+    }
+
+    @Test
+    fun generateSamplesAtGivenProbability() {
+        val sampleCount = Generator.create { it.nextInt(5, Int.MAX_VALUE) }
+            .withSamples(1, 2, 3, 4, probability = 0.5)
+            .randomSequence(0)
+            .take(1000)
+            .count { it < 5 }
+
+        assertTrue(sampleCount in 420..580)
     }
 }
 
@@ -33,7 +65,7 @@ class WithNullTest : AbstractGeneratorTest() {
         Generator.create { it.nextInt(5, Int.MAX_VALUE) }.withNull()
 
     @Test
-    fun samplesContainsNull() {
-        assertTrue(Generator.create { Any() }.withNull().samples.any { it == null })
+    fun generatesNull() {
+        assertTrue(Generator.create { Any() }.withNull().randomSequence(0).take(50).any { it == null })
     }
 }