+++ generator/stdlib/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/stdlib/FiniteValueSetGeneratorTest.kt
@@ -7,7 +7,7 @@ import kotlin.test.Test
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
 
-class FiniteValueSetGenerator : AbstractGeneratorTest() {
+class FiniteValueSetGeneratorTest : AbstractGeneratorTest() {
     override val generator: Generator<Int> = Generator.of(0, 1, 2, 3, 4)
 
     @Test
@@ -29,10 +29,5 @@ class FiniteValueSetGenerator : AbstractGeneratorTest() {
         }
 
         assertTrue(counts.all { it in 1800..2200 })
-    }
-
-    @Test
-    fun provideNoSamples() {
-        assertTrue(generator.samples.isEmpty())
     }
 }