+++ generator/stdlib/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/stdlib/ListGeneratorTest.kt
@@ -20,23 +20,23 @@ class ListGeneratorTest : AbstractGeneratorTest() {
     }
 
     @Test
-    fun samplesContainsEmpty() {
-        assertTrue(Generator.lists<Int>().samples.any { it.isEmpty() })
+    fun generatesEmpty() {
+        assertTrue(Generator.lists<Int>().randomSequence(0).take(50).any { it.isEmpty() })
     }
 
     @Test
-    fun samplesContainsSingletons() {
-        assertTrue(Generator.lists<Int>().samples.any { it.size == 1 })
+    fun generatesSingletons() {
+        assertTrue(Generator.lists<Int>().randomSequence(0).take(50).any { it.size == 1 })
     }
 
     @Test
-    fun noEmptySampleWhenMinSizeIsGreaterThan0() {
-        assertTrue(Generator.lists(Generator.ints(), minSize = 1).samples.none { it.isEmpty() })
+    fun doesNotGenerateEmptyWhenMinSizeIsGreaterThan0() {
+        assertTrue(Generator.lists(Generator.ints(), minSize = 1).randomSequence(0).take(1000).none { it.isEmpty() })
     }
 
     @Test
-    fun noSingletonSampleWhenMinSizeIsGreaterThan1() {
-        assertTrue(Generator.lists(Generator.ints(), minSize = 2).samples.none { it.size <= 1 })
+    fun doesNotGenerateSingletonWhenMinSizeIsGreaterThan1() {
+        assertTrue(Generator.lists(Generator.ints(), minSize = 2).randomSequence(0).take(1000).none { it.size <= 1 })
     }
 
     @Test
@@ -80,16 +80,6 @@ class NonEmptyListGeneratorTest : AbstractGeneratorTest() {
             .take(200)
 
         assertTrue(values.all { it.size in 1..12 })
-    }
-
-    @Test
-    fun samplesDoesNotContainsEmpty() {
-        assertTrue(Generator.nonEmptyLists<Int>().samples.none { it.isEmpty() })
-    }
-
-    @Test
-    fun samplesContainsSingletons() {
-        assertTrue(Generator.nonEmptyLists<Int>().samples.any { it.size == 1 })
     }
 
     @Test